requirements.txt
apscheduler==3.10.4
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
feedparser==6.0.11
python-dotenv==1.0.0
pytz==2023.3
